% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_regions.R
\name{bed_expand_bed_region}
\alias{bed_expand_bed_region}
\title{Title}
\usage{
bed_expand_bed_region(bed_region_size, start, end, n_reg, uncov)
}
\arguments{
\item{uncov}{}
}
\value{

}
\description{
Title
}
\examples{

}

