% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_regions.R
\name{bed_region_split}
\alias{bed_region_split}
\title{mutate function bed_region_split}
\usage{
bed_region_split(bed_region, bed_region_size)
}
\arguments{
\item{bed_region_size}{}
}
\value{

}
\description{
mutate function bed_region_split
}
\examples{

}

