% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_regions.R
\name{bedRegionSplit}
\alias{bedRegionSplit}
\title{mutate function bedRegionSplit}
\usage{
bedRegionSplit(bed_region, bed_region_size)
}
\arguments{
\item{bed_region_size}{}
}
\value{

}
\description{
mutate function bedRegionSplit
}
\examples{

}

