% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gim.R
\name{calc_gim}
\alias{calc_gim}
\title{GIM}
\usage{
calc_gim(res_df, grp = "id_bam", r_size = NULL)
}
\arguments{
\item{res_df}{results data frame}

\item{grp}{string. Variable that groups results into samples.}

\item{r_size}{integer. Region size used in computation. The defualt is \code{NULL}
in which case it is automatically calculated form \code{res_df}.}
}
\description{
Compute GIM for samples or groups of samples
}
