% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_regions.R
\name{expandBedRegion}
\alias{expandBedRegion}
\title{Title}
\usage{
expandBedRegion(bed_region_size, start, end, n_reg, uncov)
}
\arguments{
\item{uncov}{}
}
\value{

}
\description{
Title
}
\examples{

}

