% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full-genome-regions.R
\name{full_genome_regions}
\alias{full_genome_regions}
\title{Create regularly spaced regions along the whole genome.}
\usage{
full_genome_regions(genome_version, region_size = NULL, reg_overlap = 0)
}
\arguments{
\item{genome_version}{character. The genome version of your bam file.
Currently supported are \code{"hg19"} and \code{"hg38"}.}

\item{region_size}{integer. Size of the regions computed along the genome.
Should ideally be a power of 2, if it isn't it is coerced to one.}

\item{reg_overlap}{numeric. fraction of overlap for regions.}
}
\value{
A \code{tbl_df} with chromosome, start, and end columns.
}
\description{
Create regularly spaced regions along the whole genome.
}
\examples{
full_genome_regions("hg19", region_size = 2^20)

}
\seealso{
Other data import functions: \code{\link{read_bed}},
  \code{\link{read_cov}}, \code{\link{read_par_file}}
}
