% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_regions.R
\name{read_bed}
\alias{read_bed}
\title{Read bed bed_file with bed_regions to run spector}
\usage{
read_bed(bed_file, header = FALSE, ucsc_coord = FALSE,
  bed_region_size = NULL)
}
\arguments{
\item{bed_file}{Path to the bed file}

\item{header}{logical. Set \code{TRUE} if the bed file contains a header.}

\item{ucsc_coord}{logical. Set \code{TRUE} if coordinates start with 0 like
ucsc.}

\item{bed_region_size}{integer. Region size to normalise bed file.}
}
\value{
A \code{data.frame} used in the following calculations
}
\description{
Read bed bed_file with bed_regions to run spector
}
\seealso{
Other data import functions: \code{\link{read_cov}},
  \code{\link{read_par_file}}
}
