% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spector_qc.R
\name{read_par_file}
\alias{read_par_file}
\title{Read spector id file.}
\usage{
read_par_file(id_path)
}
\arguments{
\item{id_path}{path to the id file}
}
\description{
Read spector id file.
}
\seealso{
Other data import functions: \code{\link{full_genome_regions}},
  \code{\link{read_bed}}, \code{\link{read_cov}}
}
