% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spector.R
\name{spector}
\alias{spector}
\title{Run spector for a specified path or specified files}
\usage{
spector(f_bam = NULL, file_type = "list", f_delim = "\\t",
  f_head = FALSE, out_F = NULL, file_cores = 1, smr_var = "rms",
  save_out = FALSE, ...)
}
\arguments{
\item{f_bam}{path to bam file, txt file with paths or folder}

\item{file_type}{type of file passed to f_bam, options are \code{"list"},
\code{"bam"}, and \code{"dir"}}

\item{f_delim}{file delimter, only used when \code{file_type = "list"}}

\item{f_head}{binary does the \code{f_bam} file have a header,
only used when \code{file_type = "list"}}

\item{out_F}{path to output folder}

\item{file_cores}{number of cores that should be used}
}
\value{
Output files saved in folder \code{out_F}, also saves a reference
         file with file names
}
\description{
Takes in a path to a folder of .bam files, a single .bam file or a file
containing a list with the first column being a file paths and the
second column an id and the final column a groupID
}

