% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read-rigion-bed.R
\name{spector_bed}
\alias{spector_bed}
\title{Read bam file with regions}
\usage{
spector_bed(file, header = FALSE, ucsc.coord = TRUE)
}
\arguments{
\item{file}{Path to the bed file}

\item{header}{\code{TRUE} if the bed file contains a header}

\item{ucsc.coord}{\code{TRUE} if coordinates start with 0 like ucsc}
}
\value{
A \code{data.frame} used in the following calculations
}
\description{
Read bam file with regions
}

