% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spector_plot.R
\name{spector_plot}
\alias{spector_plot}
\title{Plots for SpECtOR}
\usage{
spector_plot(id_file = NULL, res_df = NULL, res_p = NULL,
  f_head = FALSE, plot_type = "boxplot", plot_var = "rms", out_F = NULL)
}
\arguments{
\item{id_file}{}

\item{res_df}{}

\item{res_p}{}

\item{f_head}{}

\item{plot_type}{}

\item{plot_var}{}

\item{out_F}{}
}
\value{
ggplot variables
}
\description{
Plots for SpECtOR
}

