% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spector_qc.R
\name{spector_qc}
\alias{spector_qc}
\title{Compute QC (spector metric) for bam files.}
\usage{
spector_qc(f_bam = NULL, file_type = "bam", f_delim = "\\t",
  out_F = NULL, file_cores = 1, smr_var = "rms", save_out = FALSE, ...)
}
\arguments{
\item{f_bam}{a string with path to \code{bam} file(s), it can link to
\code{*.bam} file (the full relative path is required), a folder with
\code{*.bam} files, or a file with with structure specified later.}

\item{file_type}{type of file passed to f_bam (Optional). This is to ensure
the automated checks pick up the correct format. The possible options
are \code{"list"}, \code{"bam"}, and \code{"dir"}.}

\item{f_delim}{the delimiter character used in \code{f_bam} file if not
\code{bam} or folder (Optional). It is only used when
\code{file_type = "list"}, with the default \code{f_delim = "\t"}.}

\item{out_F}{(Optional) Folder path to save output. If omitted results
will be returned, but not saved.}

\item{file_cores}{integer. Optional number indicating if the QC should be
computed in parallel across all input files.}

\item{smr_var}{deprecated. Variable to use to compute summary.}

\item{save_out}{logical. Indicating if output from \code{spector_qc()}
should be saved.}

\item{...}{variables passed on to downstream function, please see
\code{\link{spector_metric}()} for further details on the parameters.}
}
\value{
Output is a \code{tbl_df} object with a metric value for each region.
        Optionally the output can also be saved, if \code{out_F} is provided.
}
\description{
Wavelet based technique to compute a quality metric for regions
across the genome.
}
\details{
This is the main function to use for QC in the \pkg{spector} package. It will
compute a quality control metric for specific regions across the genome.
The default regions, supplied in the package, are based
on the genome in a bottle project
(\href{http://jimb.stanford.edu/giab/}{giab}) reliable regions, calculated
using \code{ReliableGenome} (\href{http://github.com/popitsch/wtchg-rg}{RG}).
It is also possible to supply custom regions as a bed file or a
\code{data.frame} object.

It is important to supply full paths to \code{f_bam}, and \code{f_bed}.
Though the path can be relative to the current working directory, which can
be set with \code{\link[base]{setwd}()}. This also applies to the first
column of a parameter file that can be supplied to \code{f_bam}.
}
\examples{
\dontrun{
# Compute QC on sampl1.bam with default options
spector_qc(f_bam = "sample1.bam")
}

# Compute QC on sample1.bam with custom region size
# Note. The full command is bed_region_size, but because of partial matching
# we can use this partial command.
spector_qc(f_bam = "tests/testthat/sample1.bam", region_size = 2^14)

# Compute QC on sample1.bam with custom bed file
spector_qc(f_bam = "tests/testthat/sample1.bam",
 f_bed = "tests/testthat/basic.bed")

# Compute QC and save output results
spector_qc(f_bam = "tests/testthat/sample1.bam",
 f_bed = "tests/testthat/basic.bed", out_F = "~/", save_out = T)

}

