% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{spector_samples}
\alias{spector_samples}
\title{Get path for samples provided in the spector package.}
\usage{
spector_samples(file)
}
\arguments{
\item{file}{file name}
}
\value{
path to \code{file}
}
\description{
Get path for samples provided in the spector package.
}
\examples{
spector_samples("sample1.bam")
}

