% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_file_list.R
\name{spectorList}
\alias{spectorList}
\title{Run spector for a list of bam file.}
\usage{
spectorList(fs_bam, id_v, s_v, out_F, file_cores = 1, save_out, chr_cores,
  ...)
}
\arguments{
\item{save_out}{}
}
\value{

}
\description{
Run spector for a list of bam file.
}

