% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_regions.R
\name{read_bed}
\alias{read_bed}
\title{Read bed bed_file with bed_regions to run spector}
\usage{
read_bed(bed_file, header = FALSE, ucsc_coord = FALSE,
  bed_region_size = NULL)
}
\arguments{
\item{bed_file}{Path to the bed file}

\item{header}{\code{TRUE} if the bed file contains a header}

\item{ucsc_coord}{\code{TRUE} if coordinates start with 0 like ucsc}
}
\value{
A \code{data.frame} used in the following calculations
}
\description{
Read bed bed_file with bed_regions to run spector
}

