% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read-coverage.R
\name{read_cov}
\alias{read_cov}
\title{Read the coverage of a region}
\usage{
read_cov(f.name, chr, start, end, n.read)
}
\arguments{
\item{f.name}{}

\item{chr}{}

\item{start}{}

\item{end}{}

\item{n.read}{}
}
\value{
Vector of the coverage in a given region
}
\description{
Read the coverage of a region
}

