% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-coverage.R
\name{read_cov}
\alias{read_cov}
\title{Read the coverage for regions on one chromosome.}
\usage{
read_cov(f_name, chr, start, end, n_read = NULL)
}
\arguments{
\item{f_name}{string. Full path to bam file.}

\item{chr}{string. Chromosome name, make sure it follows to naming
standards in your bam file.}

\item{start}{integer vector. Starting coordinates for regions to read.}

\item{end}{integer vector. Starting coordinates for regions to read.}

\item{n_read}{integer. total number of reads in the bam file.}
}
\value{
GRanges object with reads from a given chromosome.
}
\description{
Read the coverage for regions on one chromosome.
}
\seealso{
Other data import functions: \code{\link{full_genome_regions}},
  \code{\link{read_bed}}, \code{\link{read_par_file}}
}
