% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spector.R
\name{spector}
\alias{spector}
\title{Run spector for a specified path or specified files}
\usage{
spector(bam_f = NULL, file_type = "list", f_delim = "\\t",
  f_head = FALSE, out_F = NULL, n_core = NULL, spector_plot = TRUE,
  plot_type = "boxplot", plot_var = "rms", samtools_cmd = "samtools", ...)
}
\arguments{
\item{bam_f}{path to bam file, txt file with paths or folder}

\item{file_type}{type of file passed to bam_f, options are \code{"list"},
\code{"bam"}, and \code{"dir"}}

\item{f_delim}{file delimter, only used when \code{file_type = "list"}}

\item{f_head}{binary does the \code{bam_f} file have a header,
only used when \code{file_type = "list"}}

\item{out_F}{path to output folder}

\item{n_core}{number of cores that should be used}

\item{plot_type}{Vector for the type of plots to be saved. Options include
\code{"boxplot", "circle", "dot"}, see \code{spector_plot} for
further details.}

\item{plot_var}{The variable to be used for plotting, default is \code{"rms"}
for wavelets.}

\item{samtools_cmd}{System command for samtools including full path if necessary}

\item{...}{additional options passed to \code{gr_metric()}}

\item{r.region}{choose between the default \code{"10k"} **or** \code{"20k"}
\code{"custom"} regions}

\item{f.bed}{file path for bed file if \code{r.region = "custom"}}

\item{bed.header}{binary, \code{TRUE} if bed file has a header}
}
\value{
Output files saved in folder \code{out_F}, also saves a reference file
         with file names
}
\description{
Takes in a path to a folder of .bam files, a single .bam file or a file
containing a list with the first column being a file paths and the
second column an id and the final column a groupID
}

