% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spector.R
\docType{package}
\name{spector}
\alias{spector}
\alias{spector-package}
\title{spector: SEquence COverage Roughness}
\description{
spector provides an efficient and fast quality control for genomic data.
}
\details{
It has three main goals:

\itemize{
\item Identify regions in genomic data that are badly sequenced.
\item Determine sample quality.
\item Allow for easy comparison of quality between samples.
}

The main interface to spector quality control is \code{\link{spector_qc}()}

To learn more about spector, start with the vignettes:
\code{browseVignettes(package = "spector")}
}

