% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sr-metric.R
\name{spector_metric}
\alias{spector_metric}
\title{Calculate SpECtOR metric for one bam file}
\usage{
spector_metric(f.bam = NULL, stl_cmd = NULL, r.region = "10k",
  f.bed = NULL, bed.header = FALSE, metric = "wavelet", f.method = NA)
}
\arguments{
\item{f.bam}{}

\item{r.region}{}

\item{f.bed}{}

\item{bed.header}{}

\item{metric}{}
}
\value{
Metric for all specified regions
}
\description{
Calculate SpECtOR metric for one bam file
}

