% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr-metric.R
\name{spector_metric}
\alias{spector_metric}
\title{Calculate SpECtOR metric for one bam file}
\usage{
spector_metric(f_bam = NULL, region_size = NULL, f_bed = NULL,
  bed.header = FALSE, f.method = NA, region_giab = TRUE)
}
\arguments{
\item{region_size}{choose size of the regions defualt NULL chooses the max
power of 2 that fits in the smallest region}

\item{f_bed}{file path for bed file if \code{region_giab = FALSE}}

\item{bed.header}{binary, \code{TRUE} if bed file has a header}

\item{region_giab}{logial indicating if giab regions use or not, the default
TRUE}
}
\value{
Metric for all specified regions
}
\description{
Calculate SpECtOR metric for one bam file
}

