% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spector_plot.R
\name{spector_plot}
\alias{spector_plot}
\title{Plots for SpECtOR}
\usage{
spector_plot(id_file = NULL, res_df = NULL, res_p = NULL,
  f_head = FALSE, plot_type = "boxplot", plot_var = "rms",
  out_F = FALSE)
}
\arguments{
\item{out_F}{}
}
\value{
ggplot variables
}
\description{
Plots for SpECtOR
}

