% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_file_list.R
\name{spectorList}
\alias{spectorList}
\title{spectorList}
\usage{
spectorList(fs_bam, id_v, grp_v, s_v, out_F, n_core = 1, save_out, ...)
}
\arguments{
\item{save_out}{}
}
\value{

}
\description{
spectorList
}

